column ajust array temporalCoverage
